Started Setup for database management
After several days of tweaking and research, this is the way I found that spring boot connects to postgre. It will automatically populate tables and columns in your database as long as the database is set up in a local install of postgre. columns are entered as fields. each instance of one of these objects is a row. neat!